+++ tests/test_mock_filemanager.h
@@ -29,10 +29,10 @@ struct MockFileManager : FileManager
 {
   typedef std::shared_ptr<MockFileManager> Ptr;
 
-  MOCK_METHOD2(Open, void(std::string const& uri, unsigned long long time));
-  MOCK_METHOD2(OpenActiveChild, void(std::string const& uri, unsigned long long time));
-  MOCK_METHOD1(OpenTrash, void(unsigned long long time));
-  MOCK_METHOD1(EmptyTrash, void(unsigned long long time));
+  MOCK_METHOD2(Open, void(std::string const& uri, uint64_t time));
+  MOCK_METHOD2(OpenActiveChild, void(std::string const& uri, uint64_t time));
+  MOCK_METHOD1(OpenTrash, void(uint64_t time));
+  MOCK_METHOD1(EmptyTrash, void(uint64_t time));
   MOCK_CONST_METHOD0(OpenedLocations, std::vector<std::string>());
   MOCK_CONST_METHOD1(IsPrefixOpened, bool(std::string const& uri));
   MOCK_CONST_METHOD0(IsTrashOpened, bool());
